fix(question): fix UpdateQuestion logic and remove AutoMapper
- Remove AutoMapper from UpdateQuestion handler
- Apply manual field updates
- Add validation for exam start time
- Ensure options and correct option integrity

+++ src/ExamSystem.Application/Features/Questions/Commands/UpdateQuestion/UpdateQuestionCommand.cs
@@ -8,7 +8,7 @@ namespace ExamSystem.Application.Features.Questions.Commands.UpdateQuestion
         int ExamId,
         int QuestionId,
         string? QuestionText,
-        int? NewQuestionMark,
+        double? NewQuestionMark,
         List<UpdateOptionRequest>? Options,
         int? NewCorrectOptionId) : IRequest<Result>;
 }
